fix: split dogfood landing from app main
